+++ sdk/servicebus/service-bus/src/core/batchingReceiver.ts
@@ -36,6 +36,8 @@ export class BatchingReceiver extends MessageReceiver {
    */
   private detachedError: AmqpError | Error | undefined = undefined;
 
+  private _finalActionHandler: (() => void) | undefined = undefined;
+
   /**
    * Instantiate a new BatchingReceiver.
    *
@@ -57,6 +59,9 @@ export class BatchingReceiver extends MessageReceiver {
     // Clears the token renewal timer. Closes the link and its session if they are open.
     await this._closeLink(this._receiver);
     this.detachedError = receiverError;
+    if (this.isReceivingMessages && typeof this._finalActionHandler === "function") {
+      this._finalActionHandler();
+    }
   }
 
   /**
@@ -110,7 +115,9 @@ export class BatchingReceiver extends MessageReceiver {
       };
 
       // Final action to be performed after maxMessageCount is reached or the maxWaitTime is over.
-      const finalAction = (): void => {
+      const finalAction = (this._finalActionHandler = (): void => {
+        // clear finalActionHandler so that it can't be called multiple times.
+        this._finalActionHandler = undefined;
         if (this._newMessageReceivedTimer) {
           clearTimeout(this._newMessageReceivedTimer);
         }
@@ -125,7 +132,14 @@ export class BatchingReceiver extends MessageReceiver {
           this._receiver.session.removeListener(SessionEvents.sessionError, onSessionError);
         }
 
-        if (this.detachedError) {
+        // When receiveMode is in receiveAndDelete mode, we should return those messages to the user
+        // because they have already been removed from service bus and are safe to handle.
+        // If there haven't been any received messages, then it's safe to reject the promise
+        // so that the user knows there was an underlying issue that prevented receiving messages.
+        if (
+          this.detachedError &&
+          (this.receiveMode !== ReceiveMode.receiveAndDelete || brokeredMessages.length === 0)
+        ) {
           if (this._receiver) {
             this._receiver.removeListener(ReceiverEvents.receiverDrained, onReceiveDrain);
           }
@@ -134,7 +148,8 @@ export class BatchingReceiver extends MessageReceiver {
           return reject(err);
         }
 
-        if (this._receiver && this._receiver.credit > 0) {
+        // If the receiver has been detached, there is no need to drain.
+        if (this._receiver && this._receiver.credit > 0 && !this.detachedError) {
           log.batching(
             "[%s] Receiver '%s': Draining leftover credits(%d).",
             this._context.namespace.connectionId,
@@ -159,7 +174,7 @@ export class BatchingReceiver extends MessageReceiver {
           );
           resolve(brokeredMessages);
         }
-      };
+      });
 
       // Action to be performed on the "message" event.
       const onReceiveMessage: OnAmqpEventAsPromise = async (context: EventContext) => {
@@ -386,6 +401,8 @@ export class BatchingReceiver extends MessageReceiver {
       };
 
       if (!this.isOpen()) {
+        // clear detachedError since we are reconnecting.
+        this.detachedError = undefined;
         log.batching(
           "[%s] Receiver '%s', setting max concurrent calls to 0.",
           this._context.namespace.connectionId,